Move QuestionInput to Form components
